+++ man/roxygen/meta.R
@@ -4,6 +4,6 @@ list(
     dplyr.methods = "Other dplyr methods:",
     tidyr.method = "Other tidyr methods:",
     ggplot2.utils = "Other ggplot2 utils:",
-    hypso = "Other scales and palettes for hypsometric tints:"
+    gradients = "Other gradient scales and palettes for hypsometry:"
   )
 )