+++ tests/testthat/test-scales_fill_hypso.R
@@ -39,6 +39,30 @@ test_that("Discrete scale", {
   expect_true(all(rev(adjustcolor(mod,
     alpha.f = 0.7
   )) == mod_alpha_rev))
+
+  # Change pal
+  # Create ggplot for each pal, extract colors and check
+  # that the colors are different on each plot
+
+  allpals <- unique(hypsometric_tints_db$pal)
+
+  # Get pals
+  allpals_end <- lapply(allpals, function(x) {
+    palplot <- p + scale_fill_hypso_d(
+      palette = x
+    )
+    mod_pal <- ggplot2::layer_data(palplot)$fill
+    mod_pal
+  })
+  names(allpals_end) <- allpals
+  allpals_end <- dplyr::bind_rows(allpals_end)
+
+  length_cols <- lapply(seq_len(nrow(allpals_end)), function(x) {
+    length(unique(allpals_end[x, ]))
+  })
+  length_cols <- unlist(length_cols)
+
+  expect_true(all(length(allpals) == length_cols))
 })
 
 test_that("Continous scale", {
@@ -82,6 +106,77 @@ test_that("Continous scale", {
   expect_true(all(rev(adjustcolor(mod,
     alpha.f = 0.7
   )) == mod_alpha_rev))
+
+  # Change pal
+  # Create ggplot for each pal, extract colors and check
+  # that the colors are different on each plot
+
+  allpals <- unique(hypsometric_tints_db$pal)
+
+  # Get pals
+  allpals_end <- lapply(allpals, function(x) {
+    palplot <- p + scale_fill_hypso_c(
+      palette = x
+    )
+    mod_pal <- ggplot2::layer_data(palplot)$fill
+    mod_pal
+  })
+  names(allpals_end) <- allpals
+  allpals_end <- dplyr::bind_rows(allpals_end)
+
+  length_cols <- lapply(seq_len(nrow(allpals_end)), function(x) {
+    length(unique(allpals_end[x, ]))
+  })
+  length_cols <- unlist(length_cols)
+
+  expect_true(all(length(allpals) == length_cols))
+})
+
+test_that("Continous scale tint", {
+  d <- data.frame(x = 1:5, y = 1:5, z = 21:25, l = letters[1:5])
+
+  p <- ggplot2::ggplot(d) +
+    ggplot2::geom_col(aes(x, y, fill = z))
+
+  p_init <- p + scale_fill_hypso_c()
+
+  init <- ggplot2::layer_data(p_init)$fill
+
+  # Use tint option
+  expect_error(p + scale_fill_hypso_c(palette = "x", as_tint = TRUE))
+
+  p2 <- p + scale_fill_hypso_c(as_tint = TRUE)
+
+  mod <- ggplot2::layer_data(p2)$fill
+
+  expect_true(!any(init %in% mod))
+
+  # Reverse
+  p2_rev <- p + scale_fill_hypso_c(as_tint = TRUE, direction = -1)
+  mod_rev <- ggplot2::layer_data(p2_rev)$fill
+  expect_true(!any(mod_rev %in% mod))
+
+  # Alpha
+  p2_alpha <- p + scale_fill_hypso_c(as_tint = TRUE, alpha = 0.5)
+  mod_alpha <- ggplot2::layer_data(p2_alpha)$fill
+  expect_true(all(ggplot2::alpha(mod, .5) == mod_alpha))
+
+  # Modify limits
+
+  p3 <- p + scale_fill_hypso_c(as_tint = TRUE, limits = c(20, 26))
+  mod_lims <- ggplot2::layer_data(p3)$fill
+  expect_true(!any(mod_lims %in% mod))
+  expect_true(!any(mod_lims %in% init))
+
+  # Modify also with values
+  p4 <- p + scale_fill_hypso_c(
+    as_tint = TRUE, values = c(21, seq(22, 25, .05)),
+    limits = c(19, 27)
+  )
+  mod_values <- ggplot2::layer_data(p4)$fill
+  expect_true(!any(mod_values %in% mod_lims))
+  expect_true(!any(mod_values %in% mod))
+  expect_true(!any(mod_values %in% init))
 })
 
 test_that("Breaking scale", {
@@ -142,12 +237,96 @@ test_that("Breaking scale", {
 
   mod_alpha_rev <- ggplot2::layer_data(p4)$fill
   expect_true(length(unique(mod_alpha_rev)) == 3)
+
+  # Change pal
+  # Create ggplot for each pal, extract colors and check
+  # that the colors are different on each plot
+
+  allpals <- unique(hypsometric_tints_db$pal)
+
+  # Get pals
+  allpals_end <- lapply(allpals, function(x) {
+    palplot <- p_init + scale_fill_hypso_b(
+      palette = x
+    )
+    mod_pal <- ggplot2::layer_data(palplot)$fill
+    mod_pal
+  })
+  names(allpals_end) <- allpals
+  allpals_end <- dplyr::bind_rows(allpals_end)
+
+  length_cols <- lapply(seq_len(nrow(allpals_end)), function(x) {
+    length(unique(allpals_end[x, ]))
+  })
+  length_cols <- unlist(length_cols)
+
+  expect_true(all(length(allpals) == length_cols))
+})
+
+test_that("Breaking scale tint", {
+  d <- data.frame(x = 1:5, y = 1:5, z = 21:25, l = letters[1:5])
+
+  p <- ggplot2::ggplot(d) +
+    ggplot2::geom_col(aes(x, y, fill = z))
+
+  p_init <- p + scale_fill_hypso_b()
+
+  init <- ggplot2::layer_data(p_init)$fill
+
+  # Use tint option
+  expect_error(p + scale_fill_hypso_b(palette = "x", as_tint = TRUE))
+  p2 <- p + scale_fill_hypso_b(as_tint = TRUE)
+
+  mod <- ggplot2::layer_data(p2)$fill
+
+  expect_true(!any(init %in% mod))
+
+  # Reverse
+  p2_rev <- p + scale_fill_hypso_b(as_tint = TRUE, direction = -1)
+  mod_rev <- ggplot2::layer_data(p2_rev)$fill
+  expect_true(!any(mod_rev %in% mod))
+
+  # Alpha
+  p2_alpha <- p + scale_fill_hypso_b(as_tint = TRUE, alpha = 0.5)
+  mod_alpha <- ggplot2::layer_data(p2_alpha)$fill
+  expect_true(all(ggplot2::alpha(mod, .5) == mod_alpha))
+
+
+  # Modify limits
+
+  p3 <- p + scale_fill_hypso_b(as_tint = TRUE, limits = c(20, 26))
+  mod_lims <- ggplot2::layer_data(p3)$fill
+  expect_true(!any(mod_lims %in% mod))
+  expect_true(!any(mod_lims %in% init))
+
+  # Modify also with values
+  p4 <- p + scale_fill_hypso_b(
+    as_tint = TRUE, values = c(20, seq(22, 27, .05)),
+    limits = c(19, 27)
+  )
+  mod_values <- ggplot2::layer_data(p4)$fill
+  expect_true(!any(mod_values %in% mod_lims))
+  expect_true(!any(mod_values %in% mod))
+  expect_true(!any(mod_values %in% init))
 })
 
 test_that("Palette", {
-  a <- hypso.colors(10)
-  expect_length(a, 10)
+  expect_error(hypso.colors(20, "xx"))
 
-  # NULL
-  expect_identical(hypso.colors(0), character())
+  # Check all palettes
+  allpals <- unique(hypsometric_tints_db$pal)
+
+  # Expect character(0)
+  expect_identical(hypso.colors(0), character(0))
+
+  for (i in seq_len(length(allpals))) {
+    pal <- allpals[i]
+    colors <- hypso.colors(20, pal)
+
+    expect_identical(
+      class(colors),
+      "character"
+    )
+    expect_length(colors, 20)
+  }
 })